[OPIK-861] Moved TraceDetailPanel to shared-pages folder
