+++ apps/opik-frontend/src/components/pages-shared/traces/TraceDetailsPanel/BaseTraceDataTypeIcon.tsx
