+++ apps/opik-frontend/src/components/pages-shared/traces/TraceDetailsPanel/TagList/RemovableTag.tsx
