+++ apps/opik-frontend/src/components/pages-shared/traces/TraceDetailsPanel/TraceDataViewer/FeedbackScoreRowDeleteCell.tsx
