+++ apps/opik-frontend/src/components/pages-shared/traces/TraceDetailsPanel/TraceTreeViewer/TraceTreeViewer.module.scss
